+++ crates/til_query/src/common/logical/type_reference/bits_reference.rs
@@ -1,0 +1,7 @@
+use tydi_common::numbers::Positive;
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct BitsReference<F: Clone + PartialEq> {
+    pub bits: Positive,
+    pub field: F,
+}
